test: keep response header verification out of cookbook
